refactor: BoardService.java 수정, 폴더 구조 리펙토링
BoardService.java 수정 : getNotifyBoardListBySearchPagingCriteria, getNewsBoardListBySearchPagingCriteria 의 SearchPagingCriteria 의 categorySeq 제거
