+++ board-final-back/src/main/java/com/study/boardfinalback/domain/boards/BoardType.java
@@ -1,4 +1,4 @@
-package com.study.boardfinalback.domain;
+package com.study.boardfinalback.domain.boards;
 
 public enum BoardType {
     NOTIFY, FREE, MEMBER, NEWS
