+++ board-final-back/src/main/java/com/study/boardfinalback/domain/users/UserRole.java
@@ -1,4 +1,4 @@
-package com.study.boardfinalback.domain.user;
+package com.study.boardfinalback.domain.users;
 
 public enum UserRole {
     ROLE_MEMBER, ROLE_ADMIN
