+++ board-final-back/src/main/java/com/study/boardfinalback/domain/users/UserStatus.java
@@ -1,4 +1,4 @@
-package com.study.boardfinalback.domain.user;
+package com.study.boardfinalback.domain.users;
 
 public enum UserStatus {
     ALLOWED, BANNED
