+++ board-final-back/src/main/java/com/study/boardfinalback/service/comments/CommentQueryService.java
@@ -1,7 +1,7 @@
-package com.study.boardfinalback.service;
+package com.study.boardfinalback.service.comments;
 
 import com.study.boardfinalback.apiDto.comments.CommentWithUserDto;
-import com.study.boardfinalback.repository.CommentQueryRepository;
+import com.study.boardfinalback.repository.comments.CommentQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
