+++ board-final-back/src/main/resources/mapper/boards/BoardMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.study.boardfinalback.repository.BoardRepository">
+<mapper namespace="com.study.boardfinalback.repository.boards.BoardRepository">
 
     <select id="getBoardList" resultType="Board">
         SELECT *
