+++ board-final-back/src/main/resources/mapper/comments/CommentQueryMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.study.boardfinalback.repository.CommentQueryRepository">
+<mapper namespace="com.study.boardfinalback.repository.comments.CommentQueryRepository">
 
     <!--입력받은 board_seq을 가지는 comment의 row와 users를 join하여 return-->
     <select id="getCommentWithUserDtoListByBoardSeq" parameterType="int"
