+++ board-final-back/src/test/java/com/study/boardfinalback/repository/UserRepositoryTest.java
@@ -1,9 +1,8 @@
 package com.study.boardfinalback.repository;
 
-import com.study.boardfinalback.domain.user.User;
-import com.study.boardfinalback.domain.user.UserRole;
-import com.study.boardfinalback.domain.user.UserStatus;
-import org.junit.jupiter.api.Assertions;
+import com.study.boardfinalback.domain.users.User;
+import com.study.boardfinalback.domain.users.UserRole;
+import com.study.boardfinalback.domain.users.UserStatus;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -17,7 +16,6 @@ import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(SpringExtension.class)
 @MybatisTest
